requirements.txt
attrs==23.2.0
Automat==22.10.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
cssselect==1.2.0
defusedxml==0.7.1
exceptiongroup==1.2.1
filelock==3.15.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pluggy==1.5.0
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
PyJWT==2.9.0
pyOpenSSL==24.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
SQLAlchemy==2.0.32
tldextract==5.1.2
tomli==2.0.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
zope.interface==6.4.post2
